+++ src/routes/_authenticated/storage.report.tsx
@@ -1,9 +1,6 @@
-import { createFileRoute } from '@tanstack/react-router';
-
-function StorageReportPage() {
-  return <div>Storage report</div>;
-}
+import { createFileRoute } from '@tanstack/react-router'
+import StorageReportPage from '@/features/storage-report'
 
 export const Route = createFileRoute('/_authenticated/storage/report')({
   component: StorageReportPage,
-});
+})